requirements.txt
supabase==2.3.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit==1.32.0
